I have moved the contents of `functions.py` into `webhook_tester.py` to simplify the module. I have also updated the `tests.py` file to reflect this change.
Additionally, I replaced a deprecated `datetime.utcnow()` call with `datetime.now(UTC)`.
